pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skinny-server"
version = "0.1.0"
description = "Skinny backend API for cosmetic ingredient analysis and recommendations."
readme = "docs/01_프로젝트_개요.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.14,<2.0",
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.7,<3.0",
    "psycopg2-binary>=2.9,<3.0",
    "redis>=5.2,<6.0",
    "pinecone>=6.0,<7.0",
    "langchain>=0.3,<0.4",
    "anthropic>=0.49,<1.0",
    "python-multipart>=0.0.20,<1.0",
    "PyJWT>=2.10,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
